Extract caas operator feature tests to featuretests package
