+++ cmd/jujud/agent/unit_test.go
@@ -416,7 +416,7 @@ func (s *UnitSuite) TestChangeConfig(c *gc.C) {
 
 func (s *UnitSuite) TestWorkers(c *gc.C) {
 	coretesting.SkipIfWindowsBug(c, "lp:1610993")
-	tracker := NewEngineTracker()
+	tracker := agenttest.NewEngineTracker()
 	instrumented := TrackUnits(c, tracker, unitManifolds)
 	s.PatchValue(&unitManifolds, instrumented)
 
@@ -429,7 +429,7 @@ func (s *UnitSuite) TestWorkers(c *gc.C) {
 	go func() { c.Check(a.Run(nil), gc.IsNil) }()
 	defer func() { c.Check(a.Stop(), gc.IsNil) }()
 
-	matcher := NewWorkerMatcher(c, tracker, a.Tag().String(),
+	matcher := agenttest.NewWorkerMatcher(c, tracker, a.Tag().String(),
 		append(alwaysUnitWorkers, notMigratingUnitWorkers...))
-	WaitMatch(c, matcher.Check, coretesting.LongWait, s.BackingState.StartSync)
+	agenttest.WaitMatch(c, matcher.Check, coretesting.LongWait, s.BackingState.StartSync)
 }
