+++ cmd/jujud/agent/util_test.go
@@ -46,10 +46,9 @@ import (
 )
 
 const (
-	initialMachinePassword     = "machine-password-1234567890"
-	initialUnitPassword        = "unit-password-1234567890"
-	initialApplicationPassword = "application-password-1234567890"
-	startWorkerWait            = 250 * time.Millisecond
+	initialMachinePassword = "machine-password-1234567890"
+	initialUnitPassword    = "unit-password-1234567890"
+	startWorkerWait        = 250 * time.Millisecond
 )
 
 var fastDialOpts = api.DialOpts{
@@ -276,23 +275,6 @@ func opRecvTimeout(c *gc.C, st *state.State, opc <-chan dummy.Operation, kinds .
 			c.Fatalf("time out wating for operation")
 		}
 	}
-}
-
-type mockAgentConfig struct {
-	agent.Config
-	providerType string
-	tag          names.Tag
-}
-
-func (m *mockAgentConfig) Tag() names.Tag {
-	return m.tag
-}
-
-func (m *mockAgentConfig) Value(key string) string {
-	if key == agent.ProviderType {
-		return m.providerType
-	}
-	return ""
 }
 
 type mockLoopDeviceManager struct {
